+++ BTCPayServer.Plugins.LNbank/Data/API/WalletData.cs
@@ -18,4 +18,10 @@ public class WalletData
 
     [JsonConverter(typeof(NBitcoin.JsonConverters.DateTimeToUnixTimeConverter))]
     public DateTimeOffset CreatedAt { get; set; }
+
+    [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
+    public string LnurlPayBech32 { get; set; }
+
+    [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
+    public string LnurlPayUri { get; set; }
 }